Refactor for common files
