+++ azure-blob-storage.html
@@ -32,9 +32,7 @@
         output: 0,
         align: "right",
         icon: "blob-storage.png",
-        label: function () {
-            return "Azure Blob storage"
-        },
+        label: "Azure Blob storage",
         oneditprepare: function () {
             // Proxies
             function updateProxyOptions() {
