+++ azure-iothub-device.html
@@ -23,9 +23,7 @@
         output: 0,
         align: "right",
         icon: "iot-hub.png",
-        label: function () {
-            return "Azure IotHub device"
-        },
+        label: "Azure IotHub device",
         oneditprepare: function () {
             $("#node-input-protocol").typedInput({
                 types: [{
@@ -47,14 +45,14 @@
                 }
                 RED.tray.resize();
             }
-            
+
             if (this.proxy) {
                 $("#node-input-useProxy").prop("checked", true);
             } else {
                 $("#node-input-useProxy").prop("checked", false);
             }
             updateProxyOptions();
-            $("#node-input-useProxy").on("click", function() {
+            $("#node-input-useProxy").on("click", function () {
                 updateProxyOptions();
             });
         }
