+++ azure-iothub-device.ts
@@ -1,5 +1,6 @@
 import { Message } from "azure-iot-common";
 import { Client, DeviceClientOptions } from "azure-iot-device";
+import { ProxyNode } from "./azure-common-defs";
 import { HttpsProxyAgent, HttpsProxyAgentOptions } from "hpagent";
 import { URL } from "url";
 import * as nodered from "node-red";
@@ -8,7 +9,6 @@ import {
   AzureIotHubDeviceNodeState,
   AzureIotHubDeviceConfig,
   ProtocolModule,
-  ProxyNode,
 } from "./azure-iothub-device-def";
 
 const getProtocolModule = async function (
@@ -60,7 +60,9 @@ const getClientOptions = async function (
       return {
         mqtt: {
           ...(Object.keys(proxyConfig).length > 0 && {
-            webSocketAgent: new HttpsProxyAgent(proxyConfig as HttpsProxyAgentOptions),
+            webSocketAgent: new HttpsProxyAgent(
+              proxyConfig as HttpsProxyAgentOptions
+            ),
           }),
         },
       };
@@ -68,7 +70,9 @@ const getClientOptions = async function (
       return {
         amqp: {
           ...(Object.keys(proxyConfig).length > 0 && {
-            webSocketAgent: new HttpsProxyAgent(proxyConfig as HttpsProxyAgentOptions),
+            webSocketAgent: new HttpsProxyAgent(
+              proxyConfig as HttpsProxyAgentOptions
+            ),
           }),
         },
       };
@@ -86,7 +90,10 @@ const setup = async function (this: AzureIotHubDeviceNodeState) {
     if (msg.payload !== undefined) {
       if (msg.payload! instanceof String) {
         await this.sendMessage(msg.payload! as string);
-      } else if (msg.payload instanceof Number || msg.payload instanceof Boolean) {
+      } else if (
+        msg.payload instanceof Number ||
+        msg.payload instanceof Boolean
+      ) {
         await this.sendMessage(`${msg.payload}`);
       } else {
         await this.sendMessage(JSON.stringify(msg.payload!));
@@ -102,14 +109,19 @@ const setup = async function (this: AzureIotHubDeviceNodeState) {
       await this.client.close();
       this.log("The connection to the device was closed successfully");
     } catch (e) {
-      this.error(`An error occurred when closing the connection to the device: ${e}`);
+      this.error(
+        `An error occurred when closing the connection to the device: ${e}`
+      );
     } finally {
       done();
     }
   });
 };
 
-const sendMessage = async function (this: AzureIotHubDeviceNodeState, payload: string) {
+const sendMessage = async function (
+  this: AzureIotHubDeviceNodeState,
+  payload: string
+) {
   const message = new Message(payload);
   try {
     await this.client.sendEvent(message);
